Add color_directcolor config variable
To support 24bit colours (aka truecolors aka directcolor), there are two
parts to it:

a) We need support from ncurses.  This is a compile-time constraint
b) The terminal we are running in must support it and set up
   appropriately.  This is a run-time constraint.

The $color_directcolor config variable fetches that later run-time
constraint: yes if we have support, no if we have not.

Note about the name: TrueColor is probably the more know name for 24bit
colours, however, ncurses insists that the correct term is DirectColor.
Since our implementation heavily depends on ncurses doing the colour
stuff, we name the variable if their model.

Currently, this variable is not used anywhere.