Colors: Test on startup whether 24bit colors are supported
Set $color_directcolor according to whether ncurses thinks the terminal
supports it.  We do this by testing the COLORS variable of ncurses.

It might be that the terminal supports 24bit colors but ncurses does not
recognise that.  A reliable way to figure that out is to test against
the detected colour number of ncurses.