Colors: If convert xterm colors ("color123") to RGB
If we are using direct color convert the 256 xterm colors (defined with
e.g. `color123`) and the named colours (e.g. `brightred`) into a 24bit
colour RRGGBB.

Note: If we are using direct colours (i.e. $color_truecolor is set) then
the named colours (`brightred`) and colors color0 to color15 are mapped
to a fixed RGB value (e.g. 0xFF0000) and not the colour the terminal
defines for it.

If a user wishes to use the terminals 16 colours palette, then they
should unset $color_truecolor.