Colors: Workaround for colors #000000 to #000007
The colours #000000 to #000007 correspond to the numbers 0 to 7.  Even
when calling the *_extended_* functions of ncurses those colours
correspond to the eight default terminal colours and not to some dark
blacks.  Interestingly, colour numbers 8 to 15 (the default bright
colours of the terminal), do not show this behaviour.

As a workaround we round these eight RGB colours up to #000008 (the
darkest black we can achieve) and hope nobody notices.