set-command: Handle "set" and "set all" syntax
Handle the "set" (no arguments) and "set all" syntax which query the
value of all changed or all config variables respectively.  This syntax
is handled in the command_set_query() function.

The dump code itself is not included here in parse/set.c as it depends
on the pager functionality and thus on the pager library, which the
parse library should not.