hcache: do less work when not in use
If hcache is compiled in, but it's not actually in use, we would
still go through a lot of the motions; check for flags, stat() files
(if maildir_header_cache_verify is not explicitly turned off),
and so on. We can short-circuit a lot of these operations by simply
checking early on whether the hcache is NULL or not.

Speeds up opening a large Maildir by about 2.5% (in the specific
situation where hcache is compiled in but not enabled).