browser: fix crash in op_browser_subscribe()
Fix a segmentation fault that occurs in op_browser_subscribe() when the
list of mailboxes is empty. The macro ARRAY_GET intentionally returns
NULL when an out-of-bounds index is accessed; this value is assigned
to *ff, causing a null pointer access in mutt_str_copy().

Return FR_ERROR from op_browser_subscribe() if there are no mailboxes
in the browser.