mutt: do strict date parsing without regex
Regular expressions are a very versatile tool, but due to the
generality are not very fast. Replace the strict parser,
which deals with >98% of dates in a test Maildir here,
with a straight-up hard-coded parser.

Note that mutt_date_check_month() is now made to be case-sensitive,
since it was only ever used in contexts where the regex already
checked that the token was titlecase (so the case-insensitivity
was just extra work).

Speeds up opening a large Maildir by about 12% (compared to PCRE2).