Only add the root SRCDIR to VPATH
We don't need to add all the subdirs because their .o files are listed
including the subdir path.

This unbreaks `make clean; make test/neomutt-test; make neomutt`, which
previously failed in the last command because the `main.o` target
required by `neomutt` was resolved by `test/main.o`, for the utmost
sorrow of the linker.