Fix keymap comparison
When we compare an incoming new binding against the existing ones, we do
need to return to the caller how many characters mached, but we don't
have to retain the count found during the previous comparison. Morally,
we should just return two values here: the keymap, and the number of
chars matched.

This code is scary - I might break stuff...

Fixes #3759