Don't create empty buffers for Address mailbox/personal parts
The code base is still treating `a->personal` and `a->mailbox` as
strings, i.e., a non-NULL value means a non-empty string.

Alternatively, we could change the callers to check for `buf_is_empty`,
but we would probably still check if the pointers are valid, so at this
point I guess I'd go the cheap way and *not* have a `personal` or
`mailbox` member if it's empty.