mutt_file_sanitize_filename: actually replace characters
The documentation says "characters" but it operates on bytes. This means
that Рахунок.pdf and Ziemiańska.pdf get turned into ______________.pdf
and Ziemia__ska.pdf, which is, respectively, very useful, and offensive
to me personally.

Instead, keep the old logic for bytes <=0x7F to get rid of quotes or
slashes or whatever, but allow all other characters, since they won't
interfere with whatever mangling may happen to the path.