color_directcolor: Also set the default/initial value on startup
On startup we try to detect the terminals capability for direct color
and set value of $color_directcolor accordingly.  However, we forgot to
set its initial/default value.

Set the initial/default value for $color_directcolor like we do for all
other config variables which depend on the environment, e.g. $editor.