Allow sending an empty mail
Piping /dev/null to neomutt(1) can be useful; especially for debugging
purposes.

The original behavior, which comes from mutt(1), was to accept that.
Later, we disallowed an empty stdin, to workaround a bug in xdg-open(1)
(see issue #2138).

This commit reverts to the original behavior.  The xdg-open(1) bug
should be fixed in xdg-open(1) (and it probably has been fixed already,
because I haven't been able to reproduce it, although I haven't tested
it thoroughly).

Fixes: 331702b6a96d ("abort if run without a terminal")
Fixes: b113e9e98633 ("fix command line sending")
Link: <https://github.com/neomutt/neomutt/issues/2138>
Link: <https://github.com/neomutt/neomutt/pull/2143>
Link: <https://github.com/neomutt/neomutt/issues/2176>
Link: <https://github.com/neomutt/neomutt/pull/2177>
Closes: <https://github.com/neomutt/neomutt/issues/2243>
Link: <https://github.com/neomutt/neomutt/pull/2244>
Link: <https://github.com/neomutt/neomutt/issues/4113>
Link: <https://github.com/neomutt/neomutt/pull/4116>
Cc: <https://github.com/The-Compiler>
Cc: Austin Ray <austin@austinray.io>
Cc: Pietro Cerutti <gahr@gahr.ch>
Reported-by: <https://github.com/noahrama>
Reported-by: Alejandro Colomar <alx@kernel.org>
Co-developed-by: Alejandro Colomar <alx@kernel.org>
Signed-off-by: Alejandro Colomar <alx@kernel.org>
Co-developed-by: Richard Russon (@flatcap) <rich@flatcap.org>