mutt: -C: Enable message security in modes that by default don't enable it
neomutt(1) didn't enable message security in some cases; the most
notable one being batch mode.  This beahvior was inherited from mutt(1),
and the rationale was originally to allow using mutt(1)
non-interactively, such as in cron(8) jobs, where the keyring cannot be
unlocked easily.

However, in some other cases, the keyring can be unlocked, and it's
desirable to sign and encrypt mail.  Let's allow overriding that default
by specifying the -C flag in the command line.

This allows using neomutt(1) as a driver for git-send-email(1) to enable
encryption and signing of patches.  Here's a working configuration for
that:

In <~/.gitconfig>, add this section:

	[sendemail]
		sendmailcmd = neomutt -C -H - && true

When there are several keys for a recipient, or there is no strong valid
key for a recipient, neomutt(1) would, under some circumstances, ask the
user interactively which to use.  We now need to check if stdin is a
TTY, and if it is, ask the user as previously.  If it's not, follow the
opportunistic encryption model, that is, use a strong key (if any),
otherwise use a valid key (if any --but respect
"crypt_opportunistic_encrypt_strong_keys"--), and otherwise don't
encrypt.  If "crypt_auto_encrypt", abort the send, and report an
informative error message.

Closes: <https://github.com/neomutt/neomutt/issues/1471>
Link: <https://github.com/neomutt/neomutt/pull/1475>
Link: <https://github.com/neomutt/neomutt/pull/1476>
Link: <https://github.com/neomutt/neomutt/issues/1090>
Link: <https://marc.info/?l=mutt-dev&m=169996086410329>
Link: <https://lists.mutt.org/pipermail/mutt-dev/Week-of-Mon-20220725/001383.html>
Co-developed-by: Jenya Sovetkin <e.sovetkin@gmail.com>
Co-developed-by: Richard Russon <rich@flatcap.org>
Cc: Hadmut Danisch <hadmut@danisch.de>
Cc: Marcel Schilling <marcel.schilling@uni-luebeck.de>
Signed-off-by: Alejandro Colomar <alx@kernel.org>