envelope: manage subject/real_subj together
As it stands, subject is assigned at will throughout every interested
piece of code, and real_subj is parsed with $reply_regex and set in
three unaffiliated places.

This means that, for example, IMAP mailboxes simply do not have
$reply_regex processing. Unless you change it, then they refresh
to have it. Until you reopen and then it's lost.

Mark both fields as const and introduce a mutt_env_set_subject()
setter that sets subject and parses real_subj. This interior
mutability is bypassed only in hcache serialisation, by necessity.
rfc2047 (un)parsing needs to abuse it slightly to match the calling
convention.

This ensures that real_subj always matches subject, and so $reply_regex
works as documented instead or not at all.

Fixes: https://mailman.neomutt.org/pipermail/neomutt-devel-neomutt.org/2023-December/000897.html