ncrypt/crypt_gpgme.c: Show encryption information
Show a block of information about the encryption of a message, which
includes the recipients that can decrypt the message.

This info isn't necessarily complete: the sender could hide some
recipients, so that they are able to decrypt the message, but aren't
listed.  See gpg(1) options --recipient and --hidden-recipient.  I'm not
sure, but I suspect a malicious sender could also add fake recipients,
which are listed, but cannot really decrypt the message.

Therefore, this should be taken only as an informative block, and with a
pinch of salt.

It would be good to protect the To and Cc header fields, to provide a
guarantee to the recipient that the listed recipients match the original
recipients as intended by the sender.  See the bug #4223 linked below.

Closes: <https://github.com/neomutt/neomutt/issues/4163>
Link: <https://www.gnupg.org/documentation/manuals/gpgme.pdf#Public%20Key%20Algorithms>
Link: <https://www.gnupg.org/documentation/manuals/gpgme.pdf#Decrypt>
Link: <https://linux.codidact.com/posts/291205>
Link: <https://github.com/neomutt/neomutt/issues/4223>
Co-developed-by: Richard Russon <rich@flatcap.org>
Cc: Werner Koch <wk@gnupg.org>
Cc: наб <nabijaczleweli@nabijaczleweli.xyz>
Signed-off-by: Alejandro Colomar <alx@kernel.org>