ncrypt/: Don't introduce spurious blank lines in protected blocks
TL;DR:

These blank lines are not part of the signed and/or encrypted message.
neomutt(1) was claiming so, which was incorrect.

As наб suggested (see links below), I think we were abusing blank lines
a bit too much.

Long story:

A MIME part doesn't include the trailing blank line.  That blank line
belongs to the structure of the whole message, and doesn't belong to the
body part.

A typical PGP/MIME email looks like this:

	Date: Mon, 1 Apr 2024 21:40:04 +0200
	From: Alejandro Colomar <alx@alejandro-colomar.es>
	To: spam@alejandro-colomar.es
	Subject: After
	Message-ID: <ZgsNlP1MtLAsGBA3@debian>
	MIME-Version: 1.0
	Content-Type: multipart/signed; micalg=pgp-sha512;
		protocol="application/pgp-signature"; boundary="bHQR9+uNWFqbSKv/"
	Content-Disposition: inline

	--bHQR9+uNWFqbSKv/
	Content-Type: text/plain; protected-headers=v1; charset=utf-8
	Content-Disposition: inline
	Date: Mon, 1 Apr 2024 21:40:04 +0200
	From: Alejandro Colomar <alx@alejandro-colomar.es>
	To: spam@alejandro-colomar.es
	Subject: After

	Supposedly, migadu is fixed now.  Let's see

	--bHQR9+uNWFqbSKv/
	Content-Type: application/pgp-signature; name="signature.asc"

	-----BEGIN PGP SIGNATURE-----

	iQIzBAABCgAdFiEE6jqH8KTroDDkXfJAnowa+77/2zIFAmYLDZQACgkQnowa+77/
	2zKXvBAApT2XkOxpPRRmhjawh/axbHG3NXBldiYe9DAupE4eVJLxLnv8/tf+WWHU
	oru0vdOBaO1H8VwNi22P+aXo+29HARKawlwhVmgd2auSYzrR+UaSTSwLTN0UO8e/
	KCG2LHPMrTnZjPmlWdVRjOq44hkwes32yByjGrVz6cuwtss7XAdDmxuRGpRjHB47
	EVBstReybfepSXEhPJTB+yUfUpnVZ9G4nUBDlBODHHY8T2FPB8AMQmXAUGSIn96Y
	V33lIyiQ92gBrp3UmppfegOU+Clt18UD9tGs00UmIRlagclS9gUcUrY8CeKKMOJ5
	PYjsyqjiBLTVO58lKAjXEcAOMs0us+psxmSgzqBw52bLhxlwUDytwMUJUcQiERme
	zG6lz2lbZ45PMzrtKcm4fAxoJyQgwyf5zrQG5iyqvkmngQoo3ONlqZsZPz9GQffi
	C6uB8lmRvJM4FcPs/w2Ee6bXA3xk9ClGWryFGwdJM49nGMPXyzDx3/nc5KlOAXAn
	QPqFSy+kEuzXGJXQZcYwj0o/xNU3B+EecZH148xuAzBHpk8yQd07ZOGhwJFnUq+N
	r/cr0pr4A74t+J38Bb4vLKcWgg2pqISqX6zf/0VyYEa6o7dvMxG31gRw+z/Tv+Ln
	2WKY/wWvJHbfFuhED7uNtMfiKq8JA2ZxyKi9z/V5G780/89soG0=
	=BrzD
	-----END PGP SIGNATURE-----

	--bHQR9+uNWFqbSKv/--

Of which the main body part is precisely this:

	```
	Content-Type: text/plain; protected-headers=v1; charset=utf-8
	Content-Disposition: inline
	Date: Mon, 1 Apr 2024 21:40:04 +0200
	From: Alejandro Colomar <alx@alejandro-colomar.es>
	To: spam@alejandro-colomar.es
	Subject: After

	Supposedly, migadu is fixed now.  Let's see
	```

Note no trailing blank.  When verifying the signature, that blank line
belongs to the encapsulation boundary, and if included, it would fail
the verification.  See <https://github.com/neomutt/neomutt/issues/4229>
for that.

Then, a MIME part can be further subdivided.

Quoting RFC1521:
> Each part starts with an encapsulation boundary, and then contains a
> body part consisting of header area, a blank line, and a body area.

This RFC has been obsoleted by RFC2045, but the new one is less clear in
the wording, and probably nothing really changed, so don't worry about
it.

This says that the blank line between the headers and the body is
*not* part of the body.  This is very much like with HTTP messages.

[neo]mutt(1), however, when displaying signed/encrypted email, printed
both the separating blank and the trailing blank (or it invented a blank
line before and after the body, however it is; it might be that, since
the blank goes even before the protected subject, if there is one).
These blanks were spurious, and may lead to think that the original body
actually has those blanks.  That's a lie.  Not a big one, but a lie.

Trimming those 2 blanks makes the output more correct, faithful to the
original message as it was sent, and less consuming of non-renewable
resources (think of 24-rows here).

Closes: <https://github.com/neomutt/neomutt/issues/4242>
Link: <https://github.com/neomutt/neomutt/pull/4221#issuecomment-2048251472>
Link: <https://github.com/neomutt/neomutt/issues/4229>
Link: <https://datatracker.ietf.org/doc/html/rfc1521#section-7.2>
Link: <https://datatracker.ietf.org/doc/html/rfc2045#section-2.4>
Cc: наб <nabijaczleweli@nabijaczleweli.xyz>
Cc: Pietro Cerutti <gahr@gahr.ch>
Signed-off-by: Alejandro Colomar <alx@kernel.org>