fix: cache naming
The filename of a cache consists of two parts.

    A config variable + mailbox path

Config: `$header_cache`, `$message_cache_dir` or `$news_cache_dir`

For compatibility, the mailbox path is converted to ASCII and some
characters, such as ':' are replaced with '_'.

It's important that only the mailbox path is converted and the user's
config is left as-is.