ncrypt/crypt.c: Print a blank before the body area even if the header area is empty
Also, print the protected header marker, for the same reasons.

Quoting RFC1521:
> Each part starts with an encapsulation boundary, and then contains a
> body part consisting of header area, a blank line, and a body area.

If the user has enabled $crypt_protected_headers_read, they want to see
the entire body part, which consists of a header area, a blank, and a
body area.  If there are no header fields in the header area, that's
irrelevant to the rest of it, and a blank is still meaningful (it shows
precisely that there are 0 header fields).

So, by having the blank line even if there are no header fields, we're
being more informative to the user, and more faithful about what the
contents of the original message are.

Link: <https://datatracker.ietf.org/doc/html/rfc1521#section-7.2>
Link: <https://datatracker.ietf.org/doc/html/rfc2045#section-2.4>
Link: <https://github.com/neomutt/neomutt/issues/4242>
Link: <https://github.com/neomutt/neomutt/pull/4246>
Link: <https://github.com/neomutt/neomutt/pull/4243>
Link: <https://github.com/neomutt/neomutt/pull/4227>
Cc: наб <nabijaczleweli@nabijaczleweli.xyz>
Cc: Richard Russon <rich@flatcap.org>
Cc: Pietro Cerutti <gahr@gahr.ch>
Signed-off-by: Alejandro Colomar <alx@kernel.org>