expando: add external filter
Format strings ending in the pipe character (`|`) will be rendered, then
passed to the shell.  The output of the program will be displayed.

e.g. `set sidebar_format = "sidebar.sh '%B'|"`

The script `sidebar.sh` will be passed one argument: the name of the mailbox.
The output of the script will be displayed in the sidebar.