.clang-format: Don't special-case <stddef.h>
The commit ac6ef83f3229 ("fix: build of imap/auth_sasl.c") claims to
workaround a bug in sasl.h, which supposedly is missing an
`#include <stddef.h>` for size_t.

However, I can see the include in my system:

	alx@debian:~$ apt-file search sasl/sasl.h
	libsasl2-dev: /usr/include/sasl/sasl.h
	alx@debian:~$ grep stddef /usr/include/sasl/sasl.h
	#include <stddef.h>  /* For size_t */

And the include has been there since 2012, according to a git-blame(1)
on the source repository:

	alx@debian:/tmp/cyrus-sasl$ find | grep /sasl.h
	./include/sasl.h
	alx@debian:/tmp/cyrus-sasl$ git blame ./include/sasl.h | grep stddef
	67a188693 (Ken Murchison           2012-12-20 18:14:50 -0500  124) #include <stddef.h>  /* For size_t */
	alx@debian:/tmp/cyrus-sasl$ git show 67a188693
	commit 67a188693796a14e3a76ac603104807fbbfddfc4
	Author: Ken Murchison <murch@andrew.cmu.edu>
	Date:   Thu Dec 20 18:14:50 2012 -0500

	    sasl.h: #include <stddef.h> for size_t on NetBSD

	diff --git a/include/sasl.h b/include/sasl.h
	index fef4d510..8b8a63fb 100755
	--- a/include/sasl.h
	+++ b/include/sasl.h
	@@ -121,6 +121,8 @@
	 #ifndef SASL_H
	 #define SASL_H 1

	+#include <stddef.h>  /* For size_t */
	+
	 /* Keep in sync with win32/common.mak */
	 #define SASL_VERSION_MAJOR 2
	 #define SASL_VERSION_MINOR 1

This seems to have been a red herring.  There was probably some include
problem somewhere else, and that change accidentally fixed it.

Or maybe back then in 2017, some distributions still had an already-very
old version of <sasl/sasl.h> which still didn't have the include.
However it is, the include has been there for more than a decade now,
and we don't need to workaround any bugs.

Fixes: ac6ef83f3229 ("fix: build of imap/auth_sasl.c")
Link: <https://github.com/neomutt/neomutt/issues/4284>
Cc: Richard Russon <rich@flatcap.org>
Signed-off-by: Alejandro Colomar <alx@kernel.org>