conn: do not force username in addition to client certificate
Remove the username population code so that it's only used if it has
been set explicitly. Without this, it is not possible to use a SMTP
server expecting implicit client certificate authentication as the
username will trigger SMTP authentication which will fail if the server
doesn't support it as is common in this setup.

Upstream commit: 191b0513b43d5e603f99292faa5f8ebcc1be3823