Don't force mbox stats calculations on startup
This supersedes #4346.

Here's my current best understanding of the issue:

On startup, `main()` used the flags `MUTT_MAILBOX_CHECK_FORCE | MUTT_MAILBOX_CHECK_IMMEDIATE` in its call to `mutt_mailbox_check`.
The flag `MUTT_MAILBOX_CHECK_FORCE` was checked only in two places, inconsistently:

1. in `mutt_mailbox_check`, to call `mutt_update_num_postponed()`
https://github.com/neomutt/neomutt/blob/19653add8a9b372fd078aa98662b4d2c989445a7/mutt_mailbox.c#L174-L175

2. by the mbox backend, same as it used `MUTT_MAILBOX_CHECK_FORCE_STATS`
https://github.com/neomutt/neomutt/blob/19653add8a9b372fd078aa98662b4d2c989445a7/mbox/mbox.c#L1691

The result is that main is forcing a stats check on mbox mailboxes, even though that was never the intention.

As correctly pointed out in #4346, in
6bad3733ee2d59d04db522d396e4b999b28a2811 we inadvertently started propagating `MUTT_MAILBOX_CHECK_FORCE` from `mutt_mailbox_check` to `mailbox_check`, which initially took a boolean flag, aptly named `check_stats`.

So, because the mbox backend had a `MUTT_MAILBOX_CHECK_FORCE || MUTT_MAILBOX_CHECK_FORCE_STATS` logic, we ended up checking stats unnecessarily.

This PR clarifies the flags a bit.

* `MUTT_MAILBOX_CHECK_FORCE` becomes `MUTT_MAILBOX_CHECK_POSTPONED`, since this is what it does.
* it clarifies that `MUTT_MAILBOX_CHECK_STATS` is only used in the context of the `<check-stats>` function