Use the correct delimiter when completing an IMAP path
Fixes #2422

This turned out to be uglier than I expected.

When we expand `=` or `+` to the value of `folder` in the completion code, we insert a `/` between `folder` and what comes next, via `buf_concat_path`.
There's also a comment saying that it's safe to do so, because `imap_complete` will rewrite it:
https://github.com/neomutt/neomutt/blob/19653add8a9b372fd078aa98662b4d2c989445a7/complete/complete.c#L80-L90
Except it does now. And how could it? I suspect it might be possible for an IMAP using `.` as a delimiter to accept `/` as a regular mailbox character.

So, the solution is to call `imap_complete` with the two parts of the paths (e.g., `+` and `Archiv`), still separated.
In `imap_complete`, we check whether we can resolve an IMAP account with the first part of the path, and if we do, we then have a delimiter to use to build the full path.

As I said, it's not nice, but it seems to be working.

Comments and feedback, especially suggestions how to improve it, are welcome.