mutt_oauth2: offer to return SASL method and string
In order to use oauth tokens for IMAP/POP/SMTP connections, one needs to
wrap them in a way which depends on the method (bearer/xoauth2) and -
possily - the host and protocol. OAUTH2 aware clients typically know how
to do that, but if you want to debug using a direct connection you need
the wrapped token.

We compute the wrapped token anyways for `--test`, so offer to return it
instead of the plain access token if `--format [m]sasl` is given.
`msasl` prefixes the string with the SASL method in just the way you
would use it in a connection.

Note that the string depends on protocol (for OAUTHBEARER) and the email
identifying the mailbox (all methods). We default to `imap` and the
mailbox used for the registration but allow to specify the protocol with
a new option `--protocol` and the mailbox with the existing `--email`
option. This is useful for accessing shared mailboxes (specified with
`--email`) using your credentials (taken from the registration).