mutt_oauth2: offer to return SASL method and string
In order to use oauth tokens for IMAP/POP/SMTP connections, one needs to
wrap them in a way which depends on the method (bearer/xoauth2) and -
possily - the host and protocol. OAUTH2 aware clients typically know how
to do that, but if you want to debug using a direct connection you need
the wrapped token.

We compute the wrapped token anyways for `--test`, so offer to return it
instead of the plain access token if `--sasl <protocol>` is given.