Fix serialization of Body.content_id, bump hcache base version
In #4327, I added a new member to Body and I forgot to add it to the hcache serialization.

The change in the Body structure should be enough to invalidate the current caches, but apparently it does not.
I might try to figure out why later, but for now, let's bump the hcache version to force invalidation.