send/: Generate the Message-ID earlier
This will allow protecting the field cryptographically.  thunderbird(1)
protects the Message-ID.  And neomutt(1) already protects the
In-Reply-To and References headers, so this would be more consistent.

neomutt(1) was already capable of using a Message-ID generated earlier,
if a user enabled `edit_headers` and manually generated the field, or if
a template for the message contained the field.

After this change, the Message-ID header is generated (if it was not
already present) before opening the message in the editor, so that if
`edit_headers` is set, the user will already see the header.

Postponed messages respect the Message-ID header generated this way.  If
a postponed message didn't have the header, the header is generated when
resuming it, again before opening it in the editor.

There's still code for generating a Message-ID when actually sending, so
that if the header is removed in the editor we still send a Message-ID
header with the email.

Closes: <https://github.com/neomutt/neomutt/issues/4382>
Reviewed-by: Pietro Cerutti <gahr@gahr.ch>
Cc: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Cc: наб <nabijaczleweli@nabijaczleweli.xyz>
Signed-off-by: Alejandro Colomar <alx@kernel.org>