Don't return the result of `realpath` on failure
According to POSIX, the output argument of `realpath(3)` is undefined on
failure. The Linux man page agrees, but apparently if the failure is
caused by EACCESS or ENOENT, the resolved path is still provided as
output. This is also the behaviour on FreeBSD, which specifically says
that on failure the path that caused the issue is left in the output.

Anyway, we should not rely on anything being in the output argument on
failure, and we should leave the path argument alone.