update French translation
Replaced « sauver » by « enregistrer », which is more appropriate in
this context
Replaced « boîte mail » by « boîte aux lettres », which is the more
common translation
Started using « étiquette » for tags, which is the more common and
recommended word for this concept.
Other minor tweaks and TODO markers added for later.