notmuch: refresh the Email if the filename changes
In notmuch, there can be multiple files with the same Message-Id.
e.g. Email addressed to two different addresses that you own.

If they arrive at different times notmuch will give NeoMutt the name of
the newer file.

If NeoMutt has already parsed the file into Email/Envelope/Body, then
the arrival of the second Email will invalidate this.

When we detect a change of filename, reparse the file to be sure.

Note: The Message-Id is cached in a HashTable, so that must be preserved.