help/pager: kill magic reflow
Strip out the dynamic resizing of the Help Page. (controlled by the flags
MUTT_PAGER_RETWINCH, MUTT_PAGER_NSKIP and the fake opcode OP_REFORMAT_WINCH)

The Help Page displays three columns of text.
It picks the widths of the columns and handles clipping / wrapping.

If the user resizes the Window, the Help Page regenerates the text from
scratch and restarts the Pager.

This adds a huge amount of complexity to both the Pager and the Help Page.