help: format bound functions
Refactor the display of functions with bindings.

First collect the data in an ARRAY of HelpLine.
Next measure the width of the first and second columns.
Finally display the text compactly.

This commit changes the layout of macros.
They are now displayed over two lines to make wrapping less likely.