Fix NULL pointer dereference when calling imap_logout_all()
When idata->status is set to IMAP_FATAL, it signals the connection is
in an unusable state that can't be resolved.  While doing cleanup,
imap_close_mailbox() will leave the idata->state unchanged, and will
set idata->ctx to NULL.

So this could result in a situation where
  idata->status == IMAP_FATAL
  idata->state == IMAP_SELECTED
  idata->ctx == NULL.

Normally when idata->state == IMAP_SELECTED (or higher), Mutt assumes
a mailbox is selected and the ctx should be set to the open mailbox
CONTEXT.  However, when the status is IMAP_FATAL, mutt aborts trying
to do anything with that connection.  So there is no issue.

When exiting mutt, imap_logout_all() is called, which loops through
any unclosed connections and calls imap_logout() for each.
Unfortunately, imap_logout() was overwriting idata->status to IMAP_BYE
without checking if it was previously set to IMAP_FATAL, and then
sending a "LOGOUT" commmand.

This could result in
  idata->status == IMAP_BYE
  idata->state == IMAP_SELECTED
  idata->ctx == NULL
in which case the IMAP code assumes a ctx will be present, because status
is no longer set to IMAP_FATAL.

This commit checks for an existing IMAP_FATAL status, and skips
trying to reset the status and send "LOGOUT" in that case.

Many thanks to Roberto Ricci for reporting the bug along with a stack
trace, and working with me as I slowly found the problem.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/0ac64bf3b38d513ebe0823cdfcaa5f13718726ad