Ensure that read mails are synced when expunging deleted mails
This issue has been plaguing me for years and I can't take it anymore.
When marking mails as read and deleting mails and syncing the deleted
mails to the server the mails marked as read will not be synced to the
servers. So next time the the mailbox is opened again they appear as
unread. It's really terrible especially when having gone through
thousands of mail. Sync seen mails as well.

Fixes: #3519

Signed-off-by: Christian Brauner <brauner@kernel.org>