shrink history file on every new entry (fixes #3623, #4662)
The problem was that we only checked every `save_history` times if the
history file needed compacting. That only works for the first batch of
entries. But once we reached `save_history` number of entries we need to
compact on every new entry.

I'm not convinced that this is a good fix because compacting the history
file is a quite expensive operation that we now do on every command.
However I also don't have a better idea right now.