fix: exec auto-completion
The `exec` command takes a NeoMutt function as its argument.
To know which functions are possible, the auto-completion needs to know
which Dialog is active.

Historically, this info was obtained by dark magic.

alldialogs_get_current() looked for the top-most dialog and tried to
determine its type.  This was replaced by function that used the current
focus.

Unfortunately, the moment the user hits `:` (<enter-command>), the focus
moves to the Editor Window.

Extend `struct FileCompletionData` to take a `MuttWindow` that `exec` can use.