Consistently convert headers to LF EOL convention
When NeoMutt copies a message between folders, it might add additional headers, e.g., Content-Length.
In doing so, it terminates lines with a LF character, no matter whether the existing header lines used LF or CRLF.

There was code in place to convert CRLF to LF, but it was triggered only for address headers.

This PR makes it more consistent by converting all header lines to LF convention.